Issue error for exported pointer-to-function or pointer-to-array.
We do not support exporting a pointer-to-function or a
pointer-to-array.

We correctly issued an error when attempting to export
pointer-to-function if the function has a prototype, but not if the
function lacks a prototype.  We do a recursive export check on the
pointee type, which had a case for clang::Type::FunctionProto, but not
for clang::Type::FunctionNoProto.  This eventually caused an assertion

  slangAssert failed at frameworks/compile/slang/slang_rs_export_type.cpp:416 - 'false && "Unknown type cannot be validated"'

In the case of attempting to export pointer-to-array, we detected it
and exited some of the exporting logic early, but we never issued a
diagnostic, and we eventually got a compile time segmentation
violation.

Test: aosp_flounder-eng: slang/tests, RsTest 32bit/64bit, cts -m RenderscriptTest

Bug: 29124608
Change-Id: I8a5f83829f965de0ece4929662c288594543009f
Signed-off-by: Eliminater74 <eliminater74@gmail.com>